requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
wildlife-datasets>=0.3.0
scikit-learn>=1.3.0
finch-clust>=0.1.5
networkx>=3.1
scipy>=1.11.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
jupyter>=1.0.0
ipykernel>=6.0.0
